refactor: rename components and update font usage & modify ui
