+++ app/layout.tsx
@@ -1,24 +1,19 @@
 import type { Metadata } from "next";
-import localFont from "next/font/local";
 import "./globals.css";
 import { ThemeProvider } from "@/components/theme-provider";
 import { ModeToggle } from "@/components/theme-toggle";
 import { Analytics } from "@vercel/analytics/react";
 import dynamic from "next/dynamic";
-import ErrorBoundary from "@/components/ErrorBoundary";
+import ErrorBoundary from "@/components/error-boundary";
 import { GoogleAnalytics } from "@next/third-parties/google";
 import universities from "@/data/universities";
 import { Toaster } from "@/components/ui/toaster";
+import { Inter } from "next/font/google";
 
-const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
+const inter = Inter({
+  subsets: ["latin"],
   variable: "--font-geist-sans",
-  weight: "100 900",
-});
-const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
-  variable: "--font-geist-mono",
-  weight: "100 900",
+  weight: ["400", "500", "600"],
 });
 
 export const viewport = {
@@ -155,7 +150,7 @@ export default function RootLayout({
         <link rel="dns-prefetch" href="https://logo.clearbit.com" />
       </head>
       <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased relative overflow-auto py-5`}
+        className={`${inter.className} antialiased relative overflow-auto py-5`}
       >
         <ThemeProvider
           attribute="class"
