+++ app/page.tsx
@@ -1,4 +1,4 @@
-import { UsaUniversityCountdown } from "@/components/UsaUniversityCountdown";
+import { UsaUniversityCountdown } from "@/components/usa-university-countdown";
 
 export default function Home() {
   return <UsaUniversityCountdown />;
